backend/requirements.txt
fastapi
uvicorn
PyJWT
scikit-learn
numpy
